server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
boto3==1.28.62
python-dotenv==1.0.0
pytest==7.4.2
pytest-cov==4.1.0
requests==2.31.0